stm32cube: f2: update to Cube version v1.9.6
Update the STM32CubeF2 package from version v1.9.5 to v1.9.6.
(Package source: https://github.com/STMicroelectronics/STM32CubeF2)

Additional modifications:
- drop undocumented patch to make LL_SetFlashLatency() public in
  `stm32f0xx_ll_utils.h`; Cube provides the function officially now
  (originally introduced in 6e2d3f951472597818416cfbfff9a601bc2e60e9)

+++ stm32cube/stm32f2xx/drivers/include/stm32f2xx_hal_gpio_ex.h
@@ -190,15 +190,15 @@
 /** @defgroup GPIOEx_Get_Port_Index GPIO Get Port Index
   * @{
   */
-#define GPIO_GET_INDEX(__GPIOx__)    (uint8_t)(((__GPIOx__) == (GPIOA))? 0U :\
-                                               ((__GPIOx__) == (GPIOB))? 1U :\
-                                               ((__GPIOx__) == (GPIOC))? 2U :\
-                                               ((__GPIOx__) == (GPIOD))? 3U :\
-                                               ((__GPIOx__) == (GPIOE))? 4U :\
-                                               ((__GPIOx__) == (GPIOF))? 5U :\
-                                               ((__GPIOx__) == (GPIOG))? 6U :\
-                                               ((__GPIOx__) == (GPIOH))? 7U :\
-                                               ((__GPIOx__) == (GPIOI))? 8U : 9U)
+#define GPIO_GET_INDEX(__GPIOx__) (((__GPIOx__) == (GPIOA))? 0UL :\
+                                   ((__GPIOx__) == (GPIOB))? 1UL :\
+                                   ((__GPIOx__) == (GPIOC))? 2UL :\
+                                   ((__GPIOx__) == (GPIOD))? 3UL :\
+                                   ((__GPIOx__) == (GPIOE))? 4UL :\
+                                   ((__GPIOx__) == (GPIOF))? 5UL :\
+                                   ((__GPIOx__) == (GPIOG))? 6UL :\
+                                   ((__GPIOx__) == (GPIOH))? 7UL :\
+                                   ((__GPIOx__) == (GPIOI))? 8UL : 9UL)
 /**
   * @}
   */
